+++ src/Inttegro/PaymentMethods.cs
@@ -3,7 +3,7 @@ using System.Text.Json;
 using System.Text.Json.Nodes;
 using System.Text.Json.Serialization;
 
-namespace Inttegro.Models;
+namespace Inttegro;
 
 [JsonConverter(typeof(JsonStringEnumConverter))]
 public enum MobileMoneyNetwork